大B完结付款单功能

+++ cmd/migrate/migration/models/dynamic_shop.go
@@ -103,7 +103,7 @@ type ShopCreditLog struct {
 	DeletedAt gorm.DeletedAt `json:"-" gorm:"index;comment:删除时间"`
 	Action    string         `json:"action" gorm:"type:varchar(10);comment:操作"`
 	ShopId    int            `gorm:"index;comment:小BID"`
-	Number    int        `gorm:"comment:授信额变动数值"`
+	Number    float64        `gorm:"comment:授信额变动数值"`
 	Scene     string         ` gorm:"size:30;comment:变动场景"`
 	Desc      string         ` gorm:"size:50;comment:描述/说明"`
 	Type      int            `gorm:"type:tinyint(1);default:1;index;comment:操作类型"`
